Move everything to __init__
This will make relative imports work
and imports in unittests
