+++ Pipfile
@@ -7,21 +7,23 @@ name = "pypi"
 
 [packages]
 
-Flask = "==0.12.2"
-WTForms = "==2.1"
+flask = "==0.12.2"
+wtforms = "==2.1"
 click = "==6.7"
-Flask-SQLAlchemy = "==2.3.2"
+flask-sqlalchemy = "==2.3.2"
 itsdangerous = "==0.24"
-"Jinja2" = "==2.10"
-MarkupSafe = "==1.0"
+"jinja2" = "==2.10"
+markupsafe = "==1.0"
 passlib = "==1.7.1"
-SQLAlchemy = "==1.2.5"
-Werkzeug = "==0.14.1"
+sqlalchemy = "==1.2.5"
+werkzeug = "==0.14.1"
 
 
 [dev-packages]
 
 pytest-watch = "*"
+pytest-mock = "*"
+blinker = "*"
 
 
 [requires]
